feat(l3): add opt-in tun-over-vsock workload path
